+++ .formatter.exs
@@ -1,4 +1,5 @@
 [
   import_deps: [:stream_data],
-  inputs: ["{mix,.formatter}.exs", "{config,lib,test,benchmarks}/**/*.{ex,exs}"]
+  inputs: ["{mix,.formatter}.exs", "{config,lib,test,benchmarks}/**/*.{ex,exs}"],
+  line_length: 100
 ]